+++ packages/odata-service-writer/src/data/package.ts
@@ -1,0 +1,43 @@
+import type { Editor } from 'mem-fs-editor';
+import type { Package } from '@sap-ux/project-access';
+import { hasUI5CliV3 } from '@sap-ux/project-access';
+
+/**
+ * Update the package.json with the required middlewares.
+ *
+ * @param path path to the package.json
+ * @param fs - the memfs editor instance
+ * @param addMockServer true if the mocksever middleware needs to be added as well
+ */
+export function updatePackageJson(path: string, fs: Editor, addMockServer: boolean): void {
+    const packageJson = JSON.parse(fs.read(path)) as Package;
+    packageJson.devDependencies = packageJson.devDependencies ?? {};
+    if (!hasUI5CliV3(packageJson.devDependencies)) {
+        packageJson.ui5 = packageJson.ui5 ?? {};
+        packageJson.ui5.dependencies = packageJson.ui5.dependencies ?? [];
+        if (!packageJson.ui5.dependencies.includes('@sap/ux-ui5-tooling')) {
+            packageJson.ui5.dependencies.push('@sap/ux-ui5-tooling');
+        }
+        if (
+            addMockServer &&
+            !packageJson.ui5.dependencies.includes('@sap/ux-ui5-fe-mockserver-middleware') &&
+            !packageJson.ui5.dependencies.includes('@sap-ux/ui5-middleware-fe-mockserver')
+        ) {
+            packageJson.ui5.dependencies.push('@sap-ux/ui5-middleware-fe-mockserver');
+        }
+    }
+
+    if (!packageJson.devDependencies['@sap/ux-ui5-tooling']) {
+        packageJson.devDependencies['@sap/ux-ui5-tooling'] = '1';
+    }
+
+    if (addMockServer) {
+        if (
+            !packageJson.devDependencies['@sap/ux-ui5-fe-mockserver-middleware'] &&
+            !packageJson.devDependencies['@sap-ux/ui5-middleware-fe-mockserver']
+        ) {
+            packageJson.devDependencies['@sap-ux/ui5-middleware-fe-mockserver'] = '2';
+        }
+    }
+    fs.writeJSON(path, packageJson);
+}
